split feature-profile CLI parsing into microcrate
